iso: rename module to live
It'll encompass PXE functionality too.
